+++ tests/test_mem.py
@@ -16,30 +16,31 @@ except ImportError as e:
 
 @pytest.mark.skipif(DEPS_UNAVAILABLE, reason="Requires requests and psutil")
 def test_mem_usage():
-    """Tests that memory usage does not keep increasing"""
+    """Tests that memory usage tapers off and does not keep increasing"""
     cmd = ["python", "examples/simple_site/main.py"]
     proc = Popen(cmd, stdout=PIPE, stderr=STDOUT)
     url = "http://localhost:8888"
     try:
         time.sleep(0.2)
         total = 20000
+        half = int(total/2)
         p = psutil.Process(proc.pid)
         n = 0
-        first_mem_info = None
+        half_mem_info = None
         for i in range(total):
             r = requests.get(url)
             assert r.ok
-            n += 1
-            if n == 2000:
+            if n == 1000:
                 n = 0
                 mem_info = p.memory_info()
-                if first_mem_info is None:
-                    first_mem_info = mem_info
+                if half_mem_info is None and i >= half:
+                    half_mem_info = mem_info
                 print(f"{round(i*100/total)}% - {mem_info}")
+            n += 1
         assert (
-            first_mem_info == mem_info
-        ), "ERROR: Memory leak detected\nfirst={}\nfinal={}".format(
-            first_mem_info, mem_info
+            half_mem_info == mem_info
+        ), "ERROR: Memory leak detected\half={}\nfinal={}".format(
+            half_mem_info, mem_info
         )
         print("OK!")
     except KeyboardInterrupt: